bugfix: fix double periods in episode names
