+++ python/encoder/README.md
@@ -9,6 +9,6 @@ This python script uses ffmpeg to automate the task of determining which files a
 Pass both the folder with videos and the folder to store the transcoded files. Both folders need to exist before the script will transcode.
 
 1. Make script executable - eg.
-   `chmod +x ./encoder.py`
+   `chmod +x ./main.py`
 2. Run Script - eg.
-   `./encoder.py /in/folder/ /out/folder/`
+   `./main.py /in/folder/ /out/folder/`
